Refactoring/layers: Had separated layers.js into modules; Separated collision realisation into aggregator and subfunctions
