+++ src/layers/camera.js
@@ -1,0 +1,10 @@
+export default function createCameraLayer(cameraToDraw) {
+    return function drawCameraRect(context, fromCamera) {
+        context.strokeStyle = 'purple';
+        context.beginPath();
+        context.rect(cameraToDraw.pos.x - fromCamera.pos.x,
+                     cameraToDraw.pos.y - fromCamera.pos.y,
+                     cameraToDraw.size.x, cameraToDraw.size.y);
+        context.stroke();
+    }
+}
